+++ modules/layers/src/layers/editable-geojson-layer.js
@@ -76,6 +76,7 @@ const defaultProps = {
   onEdit: () => {},
 
   pickable: true,
+  pickingRadius: 10,
   fp64: false,
   filled: true,
   stroked: true,
@@ -263,8 +264,6 @@ export default class EditableGeoJsonLayer extends EditableLayer {
 
       modeHandler.setModeConfig(props.modeConfig);
       modeHandler.setSelectedFeatureIndexes(props.selectedFeatureIndexes);
-      modeHandler.setDeckGlContext(this.context);
-      modeHandler.setLayerId(props.id);
       this.updateTentativeFeature();
       this.updateEditHandles();
     }