+++ modules/layers/src/layers/editable-layer.js
@@ -129,7 +129,7 @@ export default class EditableLayer extends CompositeLayer {
       x: screenCoords[0],
       y: screenCoords[1],
       layerIds: [this.props.id],
-      radius: 10,
+      radius: this.props.pickingRadius || 10,
       depth: 2
     });
 
@@ -187,7 +187,7 @@ export default class EditableLayer extends CompositeLayer {
       x: screenCoords[0],
       y: screenCoords[1],
       layerIds: [this.props.id],
-      radius: 10,
+      radius: this.props.pickingRadius || 10,
       depth: 2
     });