+++ modules/layers/src/mode-handlers/mode-handler.js
@@ -37,8 +37,6 @@ export class ModeHandler {
   _modeConfig: any = null;
   _selectedFeatureIndexes: number[] = [];
   _clickSequence: Position[] = [];
-  _context: Object;
-  _layerId: string;
 
   constructor(featureCollection?: FeatureCollection) {
     if (featureCollection) {
@@ -108,14 +106,6 @@ export class ModeHandler {
 
     this._selectedFeatureIndexes = indexes;
     this._setTentativeFeature(null);
-  }
-
-  setDeckGlContext(context: Object) {
-    this._context = context;
-  }
-
-  setLayerId(layerId: string) {
-    this._layerId = layerId;
   }
 
   getClickSequence(): Position[] {